+++ tests/test_taxon.py
@@ -1,7 +1,7 @@
 from nose.tools import *
 import random
 
-from mgkit.taxon import is_ancestor, lowest_common_ancestor, NoLcaFound
+from mgkit.taxon import is_ancestor, last_common_ancestor, NoLcaFound
 
 import taxon_data
 
@@ -43,38 +43,38 @@ def test_is_ancestor3():
 
 
 @with_setup(setup=taxon_data.setup_taxon_data)
-def test_lowest_common_ancestor1():
+def test_last_common_ancestor1():
     eq_(
-        lowest_common_ancestor(taxon_data.TAXONOMY, 838, 1485),
+        last_common_ancestor(taxon_data.TAXONOMY, 838, 1485),
         2
     )
 
 
 @with_setup(setup=taxon_data.setup_taxon_data)
-def test_lowest_common_ancestor2():
+def test_last_common_ancestor2():
     eq_(
-        lowest_common_ancestor(taxon_data.TAXONOMY, 1263, 1485),
+        last_common_ancestor(taxon_data.TAXONOMY, 1263, 1485),
         186802
     )
 
 
 @with_setup(setup=taxon_data.setup_taxon_data)
-def test_lowest_common_ancestor3():
+def test_last_common_ancestor3():
     eq_(
-        lowest_common_ancestor(taxon_data.TAXONOMY, 2172, 1485),
+        last_common_ancestor(taxon_data.TAXONOMY, 2172, 1485),
         131567
     )
 
 
 @with_setup(setup=taxon_data.setup_taxon_data)
-def test_lowest_common_ancestor4():
+def test_last_common_ancestor4():
     eq_(
-        lowest_common_ancestor(taxon_data.TAXONOMY, 1385, 1485),
+        last_common_ancestor(taxon_data.TAXONOMY, 1385, 1485),
         1239
     )
 
 
 @raises(NoLcaFound)
 @with_setup(setup=taxon_data.setup_taxon_data)
-def test_lowest_common_ancestor_fail1():
-    lowest_common_ancestor(taxon_data.TAXONOMY, 256318, 1485)
+def test_last_common_ancestor_fail1():
+    last_common_ancestor(taxon_data.TAXONOMY, 256318, 1485)